Gametes Inputs correction
From now, gametes contribution was estimated as follows:
gametes = (totalWeightPre - totalWeightPost) implicitly
suggesting  that the weight loss was only due to gametes expulsion.
However, during spawning, there is also a somatic weigh loss. So, we
cannot approximated the gamete inputs like that.
So, we estimated the gamete contribution as follows:

gametes = totalGonadWeightPre- totalGonadWeightPost which provide us the
weighf of spawned gametes.